+++ CHANGES.rst
@@ -4,14 +4,17 @@ Changelog
 6.2.6 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fix change_news_type view; Taxonomy doesn't index values not present in
+  the taxonomy vocabulary, so we had lot of old values not indexed and not listed
+  as available type to change.
+  [lucabel]
 
 
 6.2.5 (2024-04-17)
 ------------------
 
 - check-servizi: fixed check on relation title.
-  [daniele] 
+  [daniele]
 
 
 6.2.4 (2024-04-16)